fix(User) : Dealer와 Player의 Name지정을 User클래스에서 할 수 있도록 수정
