+++ src/main/java/blackjack/domain/Player.java
@@ -7,15 +7,7 @@ public class Player extends User {
     boolean isStay = false;
 
     public Player(String name) {
-        super();
-        validate(name);
-        this.name = StringUtil.deleteWhiteSpaces(name);
-    }
-
-    private static void validate(String name) {
-        if (StringUtil.isBlank(name)) {
-            throw new IllegalArgumentException("빈 값을 입력하셨습니다. 플레이어의 이름을 입력해주세요.");
-        }
+        super(name);
     }
 
     public ResultType decisionGameWinOrLose(int dealerScore) {
